use the new traits for hotstore gc
