+++ blockstore/badger/blockstore.go
@@ -180,7 +180,7 @@ func (b *Blockstore) CollectGarbage() error {
 	}
 
 	for err == nil {
-		err = b.db.RunValueLogGC(0.125)
+		err = b.DB.RunValueLogGC(0.125)
 	}
 
 	if err == badger.ErrNoRewrite {
