+++ blockstore/splitstore/splitstore.go
@@ -1697,30 +1697,6 @@ func (s *SplitStore) waitForMissingRefs(markSet MarkSet) {
 	}
 }
 
-func (s *SplitStore) gcHotstore() {
-	if compact, ok := s.hot.(interface{ Compact() error }); ok {
-		log.Infof("compacting hotstore")
-		startCompact := time.Now()
-		err := compact.Compact()
-		if err != nil {
-			log.Warnf("error compacting hotstore: %s", err)
-			return
-		}
-		log.Infow("hotstore compaction done", "took", time.Since(startCompact))
-	}
-
-	if gc, ok := s.hot.(interface{ CollectGarbage() error }); ok {
-		log.Infof("garbage collecting hotstore")
-		startGC := time.Now()
-		err := gc.CollectGarbage()
-		if err != nil {
-			log.Warnf("error garbage collecting hotstore: %s", err)
-			return
-		}
-		log.Infow("hotstore garbage collection done", "took", time.Since(startGC))
-	}
-}
-
 func (s *SplitStore) setBaseEpoch(epoch abi.ChainEpoch) error {
 	s.baseEpoch = epoch
 	return s.ds.Put(baseEpochKey, epochToBytes(epoch))
